removed render logics file
