+++ src/catalog-interactive.html
@@ -63,7 +63,7 @@
 
     <script src="https://unpkg.com/aos@2.3.1/dist/aos.js"></script>
     <script type="module" src="./main.js"></script>
-    <script type="module" src="./js/catalog-render.js"></script>
+    <script type="module" src="./js/catalog-logics.js"></script>
 </body>
 
 </html>
